Unify more build configuration
